requirements.txt
Django
djangorestframework
Pillow
django-cors-headers
gunicorn